+++ backend/DTO/PaymentRequest/UpdatePaymentRequestDTO.cs
@@ -2,7 +2,7 @@
 {
     public class UpdatePaymentRequestDTO
     {
-        public double Payment {  get; set; }
+        public decimal Payment {  get; set; }
 
         public string Status { get; set; } = string.Empty;
     }
